iscsi: Enable the TPG ASAP, add portals last
RTSLib appears to be unable to remove a TPG if it hasn't been enabled,
which is less than ideal. We can get away with enabling it really early
but only adding the portals at the last step to prevent initiators from
being able to connect until we're fully up and running.