backstore: cache the storage_object when we have one
This should prevent us from walking configfs more often than we have to
if a storage object exists. We don't cache it when there isn't one,
though, as there are points in the script that we expect to go from no
objects to one being present after we've created it.