iscsi: avoid "TPG cannot be disabled" on stop
When the iscsi bringup breaks part way through, the target may not yet
be enabled when we try to delete it. RTSLib unfortunately tries to
unconditionally disable the target, which fails if it's already
disabled. Work around this by enabling the target before we delete it.

Traceback (most recent call last):
  File "/usr/lib/ocf/resource.d/rtslib/iscsi", line 27, in <module>
    ISCSITargetAgent.main()
  File "build/bdist.linux-x86_64/egg/ocf/ra.py", line 462, in main
  File "build/bdist.linux-x86_64/egg/ocf/ra.py", line 523, in execute
  File "build/bdist.linux-x86_64/egg/ocf_rtslib/iscsi.py", line 310, in stop
  File "/usr/lib/python2.7/dist-packages/rtslib/target.py", line 1084, in delete
    self.enable = False
  File "/usr/lib/python2.7/dist-packages/rtslib/target.py", line 998, in _set_enable
    raise RTSLibError("TPG cannot be disabled.")
rtslib.utils.RTSLibError: TPG cannot be disabled.